refactor : api연동 수정
